+++ src/main/java/dev/ninesliced/ui/ConfigMenuPage.java
@@ -286,8 +286,6 @@ public class ConfigMenuPage extends InteractiveCustomUIPage<ConfigMenuPage.Confi
              bindClick(events, "#AdminExplorationResetConfirmBtn", "admin_exploration_reset_confirm");
              bindClick(events, "#AdminExplorationResetCancelBtn", "admin_exploration_reset_cancel");
 
-             bindClick(events, "#HostingBannerBtn", "open_hosting_link");
-
              if (openAdminViewByDefault) {
                  ui.set("#PlayerView.Visible", false);
                  ui.set("#AdminView.Visible", true);
@@ -821,28 +819,6 @@ public class ConfigMenuPage extends InteractiveCustomUIPage<ConfigMenuPage.Confi
         refreshWorldPlayersMap(player.getWorld());
     }
 
-    private void sendHostingLink(Player player) {
-        String url = "https://zap-hosting.com/ninesliced?voucher=ninesliced";
-
-        var packetHandler = playerRef.getPacketHandler();
-        var primaryMessage = Message.raw("ZAP-Hosting Partner").color("#00aa00").bold(true);
-        var secondaryMessage = Message.raw("Click the link below to get a discount!").color("#bfcdd5");
-        var icon = new ItemStack("Deco_Trophy_Harvest", 1).toPacket();
-
-        NotificationUtil.sendNotification(
-            packetHandler,
-            primaryMessage,
-            secondaryMessage,
-            icon
-        );
-
-        Message linkMessage = Message.raw("")
-            .insert(Message.raw("[BetterMap] ").color("#93844c").bold(true))
-            .insert(Message.raw("Click here to get a discount on your game server: ").color("#bfcdd5"))
-            .insert(Message.raw(url).color("#4c9cff").link(url));
-        PlayerRefUtil.resolve(player).sendMessage(linkMessage);
-    }
-
     @Override
     public void handleDataEvent(@Nonnull Ref<EntityStore> ref, @Nonnull Store<EntityStore> store, @Nonnull ConfigEventData data) {
         Player player = store.getComponent(ref, Player.getComponentType());
@@ -992,10 +968,6 @@ public class ConfigMenuPage extends InteractiveCustomUIPage<ConfigMenuPage.Confi
                     );
                 }
                 player.getPageManager().setPage(ref, store, Page.None);
-                return;
-            }
-            case "open_hosting_link" -> {
-                sendHostingLink(player);
                 return;
             }
         }